requirements.txt
fastapi 
uvicorn 
python-jose[cryptography] 
python-multipart
python-dotenv
selenium
chromedriver-autoinstaller
requests
motor==3.7.0
pymongo==4.11.1
dnspython
PyJWT
aiohttp
